requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.47.0
wappalyzer==1.0.20
openai==1.51.0
weasyprint==65.0
python-telegram-bot==21.6
dnspython==2.7.0
celery[redis]==5.4.0
redis==5.2.1
psycopg2-binary==2.9.9
gspread==6.1.4
google-auth==2.36.0
